Expand shared upsell copies for Jetpack sites
