+++ client/my-sites/stats/stats-card-upsell/stats-upsell-copy.tsx
@@ -1,6 +1,6 @@
 import { localizeUrl } from '@automattic/i18n-utils';
 import { translate } from 'i18n-calypso';
-import { SUPPORT_URL } from '../const';
+import { SUPPORT_URL, INSIGHTS_SUPPORT_URL } from 'calypso/my-sites/stats/const';
 import {
 	STATS_FEATURE_DATE_CONTROL,
 	STATS_FEATURE_UTM_STATS,
@@ -13,7 +13,34 @@ import {
 	STAT_TYPE_TOP_AUTHORS,
 	STAT_TYPE_TOP_POSTS,
 	STAT_TYPE_VIDEO_PLAYS,
+	STAT_TYPE_EMAILS_SUMMARY,
+	STAT_TYPE_INSIGHTS_ALL_TIME_STATS,
+	STAT_TYPE_INSIGHTS_MOST_POPULAR_TIME,
+	STAT_TYPE_INSIGHTS_MOST_POPULAR_DAY,
+	STAT_TYPE_INSIGHTS_ALL_TIME_INSIGHTS,
+	STAT_TYPE_TAGS,
+	STAT_TYPE_COMMENTS,
 } from '../constants';
+
+const trafficSupportLinkWithAnchor = ( anchor: string ) => {
+	return (
+		<a
+			href={ localizeUrl( `${ SUPPORT_URL }#${ anchor }` ) }
+			target="_blank"
+			rel="noopenner noreferrer"
+		/>
+	);
+};
+
+const insightsSupportLinkWithAnchor = ( anchor: string ) => {
+	return (
+		<a
+			href={ localizeUrl( `${ INSIGHTS_SUPPORT_URL }#${ anchor }` ) }
+			target="_blank"
+			rel="noopenner noreferrer"
+		/>
+	);
+};
 
 const getUpsellCopy = ( statType: string ) => {
 	switch ( statType ) {
@@ -22,13 +49,7 @@ const getUpsellCopy = ( statType: string ) => {
 				'Learn what {{link}}external links{{/link}} your visitors click on your site to reveal their areas of interest.',
 				{
 					components: {
-						link: (
-							<a
-								href={ localizeUrl( `${ SUPPORT_URL }#clicks` ) }
-								target="_blank"
-								rel="noreferrer"
-							/>
-						),
+						link: trafficSupportLinkWithAnchor( 'clicks' ),
 					},
 				}
 			);
@@ -37,13 +58,7 @@ const getUpsellCopy = ( statType: string ) => {
 				'Discover where your {{link}}visitors are located{{/link}} and identify where your traffic is coming from.',
 				{
 					components: {
-						link: (
-							<a
-								href={ localizeUrl( `${ SUPPORT_URL }#countries` ) }
-								target="_blank"
-								rel="noreferrer"
-							/>
-						),
+						link: trafficSupportLinkWithAnchor( 'countries' ),
 					},
 				}
 			);
@@ -64,13 +79,7 @@ const getUpsellCopy = ( statType: string ) => {
 				'Find out where your {{link}}visitors come from{{/link}} to optimize your content strategy.',
 				{
 					components: {
-						link: (
-							<a
-								href={ localizeUrl( `${ SUPPORT_URL }#referrers` ) }
-								target="_blank"
-								rel="noreferrer"
-							/>
-						),
+						link: trafficSupportLinkWithAnchor( 'referrers' ),
 					},
 				}
 			);
@@ -79,13 +88,7 @@ const getUpsellCopy = ( statType: string ) => {
 				'Discover the {{link}}terms and phrases{{/link}} your visitors use to find your site.',
 				{
 					components: {
-						link: (
-							<a
-								href={ localizeUrl( `${ SUPPORT_URL }#search-terms` ) }
-								target="_blank"
-								rel="noreferrer"
-							/>
-						),
+						link: trafficSupportLinkWithAnchor( 'search-terms' ),
 					},
 				}
 			);
@@ -94,13 +97,7 @@ const getUpsellCopy = ( statType: string ) => {
 				'Identify your audience’s {{link}}favorite writers{{/link}} and perspectives.',
 				{
 					components: {
-						link: (
-							<a
-								href={ localizeUrl( `${ SUPPORT_URL }#authors` ) }
-								target="_blank"
-								rel="noreferrer"
-							/>
-						),
+						link: trafficSupportLinkWithAnchor( 'authors' ),
 					},
 				}
 			);
@@ -109,13 +106,7 @@ const getUpsellCopy = ( statType: string ) => {
 				'Discover your {{link}}post and pages{{/link}} traffic in detail and learn what content resonates the most.',
 				{
 					components: {
-						link: (
-							<a
-								href={ localizeUrl( `${ SUPPORT_URL }#posts-amp-pages` ) }
-								target="_blank"
-								rel="noreferrer"
-							/>
-						),
+						link: trafficSupportLinkWithAnchor( 'posts-amp-pages' ),
 					},
 				}
 			);
@@ -124,13 +115,7 @@ const getUpsellCopy = ( statType: string ) => {
 				'Discover your {{link}}most popular videos{{/link}} and find out how they performed.',
 				{
 					components: {
-						link: (
-							<a
-								href={ localizeUrl( `${ SUPPORT_URL }#videos` ) }
-								target="_blank"
-								rel="noreferrer"
-							/>
-						),
+						link: trafficSupportLinkWithAnchor( 'videos' ),
 					},
 				}
 			);
@@ -141,9 +126,7 @@ const getUpsellCopy = ( statType: string ) => {
 				'Generate UTM parameters and track your {{link}}campaign performance data{{/link}}.',
 				{
 					components: {
-						link: (
-							<a href={ localizeUrl( `${ SUPPORT_URL }#utm` ) } target="_blank" rel="noreferrer" />
-						),
+						link: trafficSupportLinkWithAnchor( 'utm' ),
 					},
 				}
 			);
@@ -152,16 +135,39 @@ const getUpsellCopy = ( statType: string ) => {
 				'See which {{link}}devices, browsers and OS{{/link}} your visitors are using.',
 				{
 					components: {
-						link: (
-							<a
-								href={ localizeUrl( `${ SUPPORT_URL }#devices` ) }
-								target="_blank"
-								rel="noreferrer"
-							/>
-						),
+						link: trafficSupportLinkWithAnchor( 'devices' ),
 					},
 				}
 			);
+		case STAT_TYPE_EMAILS_SUMMARY:
+			return translate(
+				'View information about {{link}}emails{{/link}} sent to your subscribers.',
+				{
+					components: {
+						link: trafficSupportLinkWithAnchor( 'emails' ),
+					},
+				}
+			);
+		case STAT_TYPE_INSIGHTS_ALL_TIME_STATS:
+			return translate( 'All-time {{link}}website insights{{/link}}.', {
+				components: {
+					link: insightsSupportLinkWithAnchor( 'all-time-highlights' ),
+				},
+			} );
+		case STAT_TYPE_INSIGHTS_MOST_POPULAR_TIME:
+			return translate( 'Best day and hour to post.' );
+		case STAT_TYPE_INSIGHTS_MOST_POPULAR_DAY:
+			return translate( 'Most popular day this year.' );
+		case STAT_TYPE_INSIGHTS_ALL_TIME_INSIGHTS:
+			return translate( 'All-time views.' );
+		case STAT_TYPE_TAGS:
+			return translate( 'Track the most viewed {{link}}tags & categories{{/link}}.', {
+				components: {
+					link: insightsSupportLinkWithAnchor( ':~:text=Tags%20,%20Categories' ),
+				},
+			} );
+		case STAT_TYPE_COMMENTS:
+			return translate( 'Top comments overview.' );
 		default:
 			return translate( 'Upgrade your plan to unlock Jetpack Stats.' );
 	}
